test(#1490): narrow the accept-list to the decoder, add the Interval arm, split units
Round-2 roborev finding 3. `ArrowShape::accepts` admitted `LargeList`,
`FixedSizeList` and `Interval(MonthDayNano)` while `arrow_rows` decoded only
`List` and had no interval arm — an accept-list broader than the decoder is a
promise the harness cannot keep, and it turns a clear early failure into a
confusing late one during value projection.

  * Lists NARROWED to `DataType::List`: neither Large nor FixedSize is more
    faithful (both are layout choices, and a Parquet LIST column reads back as
    `List`), so an export switching to one SHOULD red here.
  * `Interval(MonthDayNano)` DECODED instead of dropped: it is the FAITHFUL
    duration mapping, so rejecting it would red correct output. The arm emits
    the (months, days, nanos) triple `spelling::normalize_spelling` reconciles
    both sides onto, so no text spelling is invented and re-parsed.
  * Both files now state that they must stay in sync, and
    `every_accepted_arrow_type_is_decodable` asserts it for every accepted
    representation of every corpus scalar (both text/blob widths, both duration
    forms) rather than by convention.

Campsite-rule split (#1135): the parity file reached 1552 lines against the
~1500 test target, so the fixture-free unit coverage moved to
`issue_1490_parquet_harness_units.rs` along a subject seam (901 + 671).

Counts held: 8 tables / 40,829 cells / 2 known_gap / 1 KnownTypeGap / 49 tests.
